(PC-14258) feat(ProfileDeletion): Reorganize DeleteProfile components
