+++ src/features/profile/pages/DeleteProfile/DeleteProfileSuccess.tsx
